% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adjacentDifferences}
\alias{adjacentDifferences}
\title{Calculate adjacent differences in a vector}
\usage{
adjacentDifferences(x)
}
\arguments{
\item{x}{Armadillo vector with numeric element}
}
\value{
Armadillo vector with the adjacent differences
}
\description{
All differences between adjacent elements in a vector
}
\examples{
\dontrun{
# simple call: All adjacent elements has the same length 1
adjacentDifferences(c(1,2,3,4,5))
# simple call: There are positive and negative diffferences
adjacentDifferences(c(1,2,0,14,-2))
}
}
\author{
Roberto Maestre
}
