% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{apodization}
\alias{apodization}
\title{Apodization}
\usage{
apodization(frequences, filter)
}
\arguments{
\item{frequences}{Armadillo vector with frequences}

\item{filter}{A string with a specific filter (bartlett, blackman,
connes, cosine, gaussian, hamming, hanning, welch or uniform by-default)}
}
\value{
The apoditazed frequence vector
}
\description{
Apodization, tapering or window function is a function 
used to smoothly bring a sampled signal down to zero at the edges 
of the sampled region
}
\examples{
\dontrun{
# simple call:
apodization(c(1,2,3,4), "gaussian")
apodization(c(1,2,3,4), "blackman")

}
}
\author{
Roberto Maestre
}
