% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{apodizationFt}
\alias{apodizationFt}
\title{Apodization and FT}
\usage{
apodizationFt(frequences, filter)
}
\arguments{
\item{frequences}{Armadillo vector with frequences}

\item{filter}{A string with a specific filter (bartlett, blackman,
connes, cosine, gaussian, hamming, hanning, welch or uniform by-default)}
}
\value{
A list with amplitudes, frequences, inverse frecuences and the
power spectrum calculated.
}
\description{
Apodization of a given vector with one filter and
Fourier Transform calculation
}
\examples{
\dontrun{
# simple call:
r <- ft(sin(seq(1:100)), "uniform")
plot(r$f, r$powerSpectrum)
}
}
\author{
Roberto Maestre
}
