% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateRange}
\alias{calculateRange}
\title{Split elements in ranges}
\usage{
calculateRange(nElements, nFrequencies)
}
\arguments{
\item{nElements}{Number of elements to be splitted}

\item{numFrequencies}{Number of total frequences}
}
\value{
A vector of ranges in integer format
}
\description{
Inspired in the quantmod::findPeaks() R algorithm, 
this function find the highest peaks in a time series.
}
\examples{
\dontrun{
# simple call:
calculateRange(100,100)
calculateRange(100,100)
calculateRange(10,200)
calculateRange(1,20)
}
}
\author{
Roberto Maestre
}
