% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateSpectrum}
\alias{calculateSpectrum}
\title{Calculate the spectrum from a given vector}
\usage{
calculateSpectrum(time, x)
}
\arguments{
\item{time}{Armadillo vector representing the meassurements time}

\item{x}{Armadillo vector with numeric elements}
}
\value{
A complex vector with the calculated spectrum
}
\description{
Calculate frequency, amplitude and phase from a given vector
using the Fast Fourier Tramsform algorithm.
}
\examples{
\dontrun{
# simple call:
sample <- sin(seq(1:100))
calculateSpectrum(seq(1:length(sample)),sample)
}
}
\author{
Roberto Maestre
}
