% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crosscorrelation}
\alias{crosscorrelation}
\title{Crosscorrelation}
\usage{
crosscorrelation(frequencies, lagMax = NULL, type = "correlation",
  plot = FALSE)
}
\arguments{
\item{frequencies}{The frequences to be processed}

\item{lagMax}{Maximum lag at which to calculate the acf
(null by default)}

\item{type}{character string giving the type of acf to be 
computed. Allowed values are "correlation" or 
"covariance" (correlation by default).}
}
\value{
plot If true the acf is plotted (true by default)
}
\description{
This function calculates the crosscorrelation of the
frequency vector by using a sum of gaussians
}
\examples{
\dontrun{
# simple call:
crosscorrelation(sin(c(1,2,3,4,5,4,3,2,1)), lagMax = 50, plot=T)
}
}
\author{
Roberto Maestre
}
