% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{differences}
\alias{differences}
\title{Triangular Superior absolute differences of a vector}
\usage{
differences(frequences)
}
\arguments{
\item{frequences}{Armadillo vector with frequences}
}
\value{
The vector with all differences
}
\description{
Given a vector, this function calculates the triangular
superior differences combination. All differences are
absolute and zero-difference will be drop
}
\examples{
\dontrun{
# simple call:
differences(c(1,2,3,4,5))
# simple call returning an empty vector because all 
# differences are equal to 0:
differences(rep(1,10))

}
}
\author{
Roberto Maestre
}
