% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diffHistogram}
\alias{diffHistogram}
\title{Histogram of differences}
\usage{
diffHistogram(frequences, dnu)
}
\arguments{
\item{frequences}{Armadillo vector with frequences}

\item{dnu}{Numeric value for the spectral resolution}
}
\value{
A list with bins and values representing an histogram
}
\description{
Given a vector, this function calculates the histogram
of the non-zero and absolute differences of a given vector
taking into account a Dnu
}
\examples{
\dontrun{
# simple call:
differences(c(1,2,3,4,5))
# simple call returning an empty vector because all 
# differences are equal to 0:
differences(rep(1,10))

}
}
\author{
Roberto Maestre
}
