% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findPeaks}
\alias{findPeaks}
\title{Find highest peaks in a time series}
\usage{
findPeaks(x)
}
\arguments{
\item{x}{Armadillo vector with numeric element}
}
\value{
Armadillo unsigned vector with the index representing 
the position of the selected peaks
}
\description{
Inspired in the quantmod::findPeaks() R algorithm, 
this function find the highest peaks in a time series.
}
\examples{
\dontrun{
# simple call: Returns position 1,4 and 7 with the highest peaks
findPeaks(c(1,2,1,2,10,1,2,3,1))
}
}
\author{
Roberto Maestre
}
