% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seqIntegers}
\alias{seqIntegers}
\title{Create a vector as sequence of integers}
\usage{
seqIntegers(first, last)
}
\arguments{
\item{first}{The first element of the sequence}

\item{last}{The last element of the sequence}
}
\value{
A numeric vector with the whole secuence of integers
}
\description{
This function helps to create a vector with a secuence
of integers equally separated
}
\examples{
\dontrun{
# simple call:
seqIntegers(1,10)
# simple call with negative numbers
seqIntegers(-10,10)
}
}
\author{
Roberto Maestre
}
