% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vectorRev}
\alias{vectorRev}
\title{Revert the order of an input vector}
\usage{
vectorRev(vector)
}
\arguments{
\item{vector}{Numerix vector to be reversed}
}
\value{
A numeric vector with elements in reverse mode
}
\description{
Revert the order of the elements an input vector
}
\examples{
\dontrun{
# simple call:
vectorRev(c(1,2,3,4,5,6))
# simple call with negative nummbers
vectorRev(c(1,-2,3,-4,5,-6))
}
}
\author{
Roberto Maestre
}
